added timestamp for graphs

+++ pycrunch_tracer/events/base_event.py
@@ -4,4 +4,4 @@ from abc import ABC, abstractmethod
 class Event(ABC):
     # timestamp of event
     event_name: str
-    ts: int
+    ts: float
